API Refactor/requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
Pillow==8.3.1
pkg-resources==0.0.0
pytz==2021.1
pyzbar==0.1.8
six==1.16.0
Werkzeug==2.0.1
